challenge: testing error when renaming CodeMonkeySO to CMSO
public static CMSO GetCodeMonkeyFreeSO method looks codeMonkeyFreeSO value, which initially is null.
A string[] cmguidarray is created by calling FindAssets(nameof(CMSO));
As the previous SO had a different name, it won't be found.
Solution: change the name of the new asset to match exactly as the SO class, in this case, CMSO.